requirements.txt
aiogram==3.15.0
pillow==11.0.0
python-dotenv==1.0.1
emoji==1.7.0